python/requirements.txt
# Zenoh Python SDK
eclipse-zenoh>=1.0.0

# CDR serialization
pycdr2>=1.0.0
